feat(scheme/alert): make to ignore duplicated alerts
