+++ internal/scheme/source.go
@@ -12,7 +12,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -26,38 +25,6 @@ var (
 	ErrInvalidSourceURL = errors.New("invalid source URL")
 	ErrMissingFile      = errors.New("missing file")
 )
-
-// urlSet is a set of URL.
-type urlSet []*url.URL
-
-func (s urlSet) search(u *url.URL) int {
-	return sort.Search(len(s), func(i int) bool {
-		return strings.Compare(s[i].String(), u.String()) <= 0
-	})
-}
-
-// Has check if the URL is in this urlSet or not.
-func (s urlSet) Has(u *url.URL) bool {
-	i := s.search(u)
-	if len(s) == i {
-		return false
-	}
-
-	return s[i].String() == u.String()
-}
-
-// Add adds a URL to urlSet.
-// If the URL is already added, it will be ignored.
-func (s *urlSet) Add(u *url.URL) {
-	i := s.search(u)
-	if len(*s) == i {
-		*s = append(*s, u)
-	}
-
-	if (*s)[i].String() != u.String() {
-		*s = append(append((*s)[:i], u), (*s)[i:]...)
-	}
-}
 
 func normalizeSourceURL(u *url.URL) (*url.URL, error) {
 	switch u.Scheme {
